# api.yaml
swagger: '2.0'
info:
  title: Reshigo API
  version: "1.0"
# the domain of the service
host: server.reshigo.ru
...
schemes:
  - http
# will be prefixed to all paths
produces:
  - application/json
securityDefinitions:
  basicAuth:
    type: basic
    description: Http Basic Authentication. Works over `HTTP`
paths:
...
  /user:
    get:
      description: Get user info
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'
        default:
          description: Unexpected error
    put:
      description: Update user\'s info and education and degree fields
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: user
          in: body
          description: Only some of info, education, degree, iosReview, androidReview fields required. Max length is 1000 bytes for each string field
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        200:
          description: OK
        400:
          description: Wrong data in request.
  /user/registration:
    post:
      description: Registration of a consumer. If consumer is registered, but not enabled, you can resend registration to fix phone number. It is necessary to resend code.
      tags:
        - Consumer
      parameters:
        - name: user
          in: body
          description: User to add into the system
          required: true
          schema:
            $ref: '#/definitions/Registration'
      responses:
        201:
          description: Created
        400:
          description: Parameter was not set correctly
          schema:
            $ref: '#/definitions/HttpResponseEntity'
        409:
          description: User is already registered or duplicate phone/email
          schema:
              $ref: '#/definitions/HttpResponseEntity'
        503:
          description: Problems with sms service. Maybe your number is not correct. Contact support
          schema:
              $ref: '#/definitions/HttpResponseEntity'
  /user/registration/verification:
    post:
      description: Verify registration by 4-digit code. 3 attempts maximum.
      tags:
        - Consumer
      parameters:
        - name: code
          in: body
          description: 4-digit code from sms
          required: true
          schema:
            type: string
        - name: name
          in: query
          description: user\'s name
          required: true
          type: string
      responses:
        200:
          description: OK
        403:
          description: Code is not valid
  /user/registration/verification/code:
    get:
      description: Resend user\'s verification code
      tags:
        - Consumer
      parameters:
        - name: name
          in: query
          description: user\'s name
          required: true
          type: string
      responses:
        200:
          description: OK
        503:
            description: Problems with sms service. Maybe your number is not correct. Contact support
  /user/fcm:
    put:
      description: update gcm notification token
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: token
          in: body
          description: Update firebase notification token
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
  /user/logout:
      get:
        description: clean fcm token
        tags:
          - Consumer
          - Solver
        security:
          - basicAuth: []
        responses:
          200:
            description: OK
  /user/funds:
    get:
      description: Add some funds for consumer
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: amount
          in: query
          required: false
          type: integer
          format: int64
      responses:
        200:
          description: OK
          schema:
            type: string
            description: html page for user to complete payment
        403:
          description: Amount is less than 50 or this user is solver.
        503:
          description: service unavailable, contact support
  /user/payments:
    get:
      description: Get list of payments
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Payment'
  /user/promocode:
    post:
      description: set user\'s registration promocode
      tags:
        - Consumer
      security:
        - basicAuth: []
      parameters:
        - name: promocode
          in: body
          description: characters sequence
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
        403:
          description: promocode was already set or user inserts his own code
        404:
          description: such promocode was not found
        503:
          description: service unavailable, contact support
  /user/photo:
    post:
      description: |
          Update profile\'s photo. Max size is 50 kilobytes. Use "Content-type: multipart/form-data" and put binary data of image in field "file"
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      responses:
        200:
          description: OK
        413:
          description: payload is larger than 50 kilobytes.
        500:
          description: internal error
  /user/{name}:
    get:
      description: |
        Get user\'s information. Only name, info, education and rating presented.
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
        404:
          description: user not found
  /user/{name}/photo:
    get:
      description: |
        Get user\'s photo
      tags:
        - Consumer
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
        404:
          description: user or photo not found
        500:
          description: internal error
...
definitions:
  HttpResponseEntity:
    type: object
    properties:
      error:
        type: string
        description: Description of an occured error. Every object below extends HttpResponse entity and
          in case of an error, may have not empty field to show for user
...
  User:
    type: object
    properties:
      name:
        type: string
        description: should satisfy regexp [a-zA-Z0-9_/\\\[\]\.\{\}\-]+ and can't consist only from numbers
      email:
        type: string
        description: DEPRECATED
      phone:
        type: string
      fcmToken:
        description: token to use for notifications server
        type: string
      isCustomer:
        type: boolean
      promocode:
        description: sequence of characters to invite friends
        type: string
      funds:
        type: number
        format: double
      reservedFunds:
        type: number
        format: double
      commission:
        type: number
        format: float
      info:
        type: string
        format: binary
        description: utf-8 string
      education:
        type: string
        description: utf-8 string
      degree:
        type: string
        description: utf-8 string, representing class, course or other academic degree
      campaign:
        type: string
        description: source, from where this user came
      iosReview:
        type: string
        description: if true, then user reviewed application in AppStore. can be changed via PUT /user method
      androidReview:
        type: string
        description: if true, then user reviewed application in PlayStore. can be changed via PUT /user method
  Registration:
    type: object
    properties:
      name:
        type: string
        description: should satisfy regexp [a-zA-Z0-9_/\\\[\]\.\{\}\-]+
      password:
        type: string
        description: should satisfy regexp [a-zA-Z0-9_/\\\[\]\.\{\}\-]+
      email:
        type: string
      phone:
        type: string
      token:
        type: string
      campaign:
        type: string
        description: source, from where this user came
...
  Payment:
    type: object
    description: payment info
    properties:
      id:
        type: integer
        format: int64
      amount:
        type: number
        format: float
      commission:
        type: number
        format: float
        description: Value between 0 and 1. Only required when amount if less than zero.